# specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PetCare AI - Pet Operating System",
    "description": "AI-powered platform for pet health tracking, vet collaboration, and emergency response",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/vet/patients": {
      "get": {
        "tags": [
          "Vet Dashboard"
        ],
        "summary": "Get Patients",
        "operationId": "get_patients_api_vet_patients_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/vet/patients/{pet_id}/summary": {
      "get": {
        "tags": [
          "Vet Dashboard"
        ],
        "summary": "Patient Summary",
        "operationId": "patient_summary_api_vet_patients__pet_id__summary_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/vet/appointments": {
      "get": {
        "tags": [
          "Vet Dashboard"
        ],
        "summary": "Vet Appointments",
        "operationId": "vet_appointments_api_vet_appointments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/vet/dashboard-stats": {
      "get": {
        "tags": [
          "Vet Dashboard"
        ],
        "summary": "Dashboard Stats",
        "operationId": "dashboard_stats_api_vet_dashboard_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/vet-chat/rooms": {
      "get": {
        "tags": [
          "Vet Chat"
        ],
        "summary": "List Rooms",
        "operationId": "list_rooms_api_vet_chat_rooms_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Vet Chat"
        ],
        "summary": "Create Room",
        "operationId": "create_room_api_vet_chat_rooms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/vet-chat/rooms/{room_id}/messages": {
      "get": {
        "tags": [
          "Vet Chat"
        ],
        "summary": "Get Messages",
        "operationId": "get_messages_api_vet_chat_rooms__room_id__messages_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Room Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/vet-chat/vets": {
      "get": {
        "tags": [
          "Vet Chat"
        ],
        "summary": "List Vets",
        "operationId": "list_vets_api_vet_chat_vets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRoomRequest": {
        "properties": {
          "vet_id": {
            "type": "integer",
            "title": "Vet Id"
          },
          "pet_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pet Id"
          }
        },
        "type": "object",
        "required": [
          "vet_id"
        ],
        "title": "CreateRoomRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}